requirements.txt
markitdown[docx]
openai
python-dotenv
pypandoc_binary
pdf2image
pillow
dashscope